use IPython's display in the patch as well
